backend/requirements.txt
fastapi 
uvicorn[standard] 
sqlalchemy 
beautifulsoup4 
requests 
pydantic 
langchain-core 
langchain-community 
python-dotenv 
langchain-google-genai
pymysql